+++ example/tasks/lib/features/tasks/domain/task_project.dart
@@ -12,7 +12,4 @@ abstract class TaskProject
     implements OwnedBy<TaskProject, Account>, SoftDeletable, Ordered {
   @Persisted(minLength: 1, maxLength: 80, conflict: ConflictStrategy.localWins)
   abstract final String title;
-
-  @Action()
-  Future<void> rename({required String title});
 }